Remove dependecy on `git`
Retrieving the version of `grafana-dashboard-manager` for the
purpose of logging should not require that you are running
inside of the git repository, or require you to run via `poetry`
to succeed.

I have switched to using `importlib.metadata` from the stdlib to
retrieve the package metadata which contains the version number.

Signed-off-by: Christian Witts <cwitts@gmail.com>